refactor: 컴포넌트 관련 controller는 해당 컴포넌트내의 controller 폴더로 이동
