+++ src/components/error/index.ts
@@ -1,7 +1,15 @@
+import ErrorViewController from './controller/ErrorViewController';
 import ErrorBox from './ErrorBox';
 import ErrorImg from './ErrorImg';
 import ErrorMessage from './ErrorMessage';
 import ErrorView from './ErrorView';
 import NullElementError from './NullElementError';
 
-export { ErrorBox, ErrorImg, ErrorView, ErrorMessage, NullElementError };
+export {
+  ErrorBox,
+  ErrorImg,
+  ErrorView,
+  ErrorViewController,
+  ErrorMessage,
+  NullElementError,
+};
