+++ src/components/index.ts
@@ -1,11 +1,11 @@
 import Header from './Header';
 import Label from './Label';
 import RefreshButton from './RefreshButton';
-import SearchBox from './SearchBox';
 
 export * from './error';
 export * from './modal';
 export * from './movie';
+export * from './searching';
 export * from './skeleton';
 
-export { Header, Label, RefreshButton, SearchBox };
+export { Header, Label, RefreshButton };
