+++ src/components/modal/MovieInfoModal.ts
@@ -1,9 +1,9 @@
 import { MOVIE_INFO_COMMON_CLASS } from '../../constants';
-import { ModalContainerController } from '../../controller';
 import { MovieInfo } from '../../type/movie';
 import { createElementWithAttribute } from '../../utils';
 import { MovieImg, MovieScore, MovieTitle, UserScore } from '../movie';
 
+import ModalContainerController from './controller/ModalContainerController';
 import ModalCloseButton from './ModalCloseButton';
 import ModalContainer from './ModalContainer';
