+++ src/components/modal/index.ts
@@ -1,4 +1,5 @@
 import AlertModal from './AlertModal';
+import ModalContainerController from './controller/ModalContainerController';
 import ModalCloseButton from './ModalCloseButton';
 import ModalContainer from './ModalContainer';
 import MovieInfoModal from './MovieInfoModal';
@@ -7,6 +8,7 @@ import ToastModal from './ToastModal';
 export {
   AlertModal,
   ModalContainer,
+  ModalContainerController,
   ModalCloseButton,
   MovieInfoModal,
   ToastModal,
