+++ src/components/searching/index.ts
@@ -1,0 +1,4 @@
+import SearchBox from './SearchBox';
+import SearchBoxResponsiveHandler from './SearchBoxResponsiveHandler';
+
+export { SearchBox, SearchBoxResponsiveHandler };
