+++ src/components/skeleton/index.ts
@@ -1,3 +1,4 @@
+import SkeletonController from './controller/SkeletonController';
 import SkeletonCard from './SkeletonCard';
 import SkeletonInfo from './SkeletonInfo';
 import SkeletonList from './SkeletonList';
@@ -6,6 +7,7 @@ import SkeletonTextBox from './SkeletonTextBox';
 
 export {
   SkeletonCard,
+  SkeletonController,
   SkeletonInfo,
   SkeletonList,
   SkeletonListContainer,
